Move routes to own directory
for better future routes management
